% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.toolbox.R
\name{align.fields.with.masterfield}
\alias{align.fields.with.masterfield}
\title{Segment a group of fields associated to an unit (fields related to word, or to morphem) in toolbox}
\usage{
align.fields.with.masterfield(longformat, fields = c("mb", "ge", "ps"),
  masterfield = "mb", unit_id_name = "morphems_id", included.in = NULL,
  included.in_id = "words_id")
}
\arguments{
\item{longformat}{the data table of the toolbox fields in long format (melted)}

\item{fields}{character vector: the aligned fields to tokenize (for instance, "mb", "ge" and "ps")}

\item{masterfield}{the master in the fields to be tokenized (for instance, "mb").}

\item{unit_id_name}{character vector (length-1): the name of the column to be created containing an ID for each token.}

\item{included.in}{character vector: an optional super-ordinated unit (for instance words for morphems).}

\item{included.in_idcharacter}{vector (length-1): the name of the column to be created containing an ID of the outer unit.}
}
\value{
a data frame with one token by row
}
\description{
Segment into tokens a master field (such as tx for words, or mb for morphems), and segment other fields
associated with it (such as ge", or "ps" for the "mb" master field.)
}
