% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.emeld.R
\name{concatenate.items.in.element}
\alias{concatenate.items.in.element}
\title{Same as items.in.element, but if multiple instnaces of the same field with the same
language are found, they are concatenated (with "; "). Used only for "note" in sentence.}
\usage{
concatenate.items.in.element(elements, fields, languages)
}
\arguments{
\item{elements}{see items.in.element}

\item{fields}{see items.in.element}

\item{languages}{see items.in.element}
}
\value{
a data.frame. see items.in.element.
}
\description{
Same as items.in.element, but if multiple instnaces of the same field with the same
language are found, they are concatenated (with "; "). Used only for "note" in sentence.
}
