% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.lift.R
\name{get.fields}
\alias{get.fields}
\title{Extract values according to the specification of a field in a LIFT dictionary}
\usage{
get.fields(nodes, spec, vernacular.languages, analysis.languages)
}
\arguments{
\item{nodes}{the context node to evaluate the XPath queries with}

\item{spec}{A specification: contains information in order to build one or serveral XPath queries}

\item{vernacular.languages}{character vector: the vernacular languages codes used in the dictionary}

\item{analysis.languages}{character vector: the analysis languages codes used in the dictionary}
}
\description{
Extract values according to the specification of a field in a LIFT dictionary
}
