% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.emeld.R
\name{get.languages}
\alias{get.languages}
\title{Extract the vernacular of analysis language of a EMELD document}
\usage{
get.languages(corpusdoc, type)
}
\arguments{
\item{corpusdoc}{the XML document (xml2 library)}

\item{type}{length-1 character vector, "vernacular" or "analysis"}
}
\value{
a character vector of language code.
}
\description{
Extract the vernacular of analysis language of a EMELD document
}
