% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlineaR.R
\docType{package}
\name{interlineaR-package}
\alias{interlineaR}
\alias{interlineaR-package}
\title{Parse XML document containing interlinearized texts}
\description{
Parse XML document containing interlinearized texts
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.r-project.org}
}

}
