% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interlineaR-package.R
\docType{package}
\name{interlineaR}
\alias{interlineaR}
\alias{interlineaR-package}
\alias{interlineaR-package}
\title{Importing interlinearized corpora and dictionaries as produced by descriptive linguistics softwares}
\description{
Importing interlinearized corpora and dictionaries as produced by descriptive linguistics softwares
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sylvainloiseau/interlineaR}
}

}
\author{
\strong{Maintainer}: Sylvain Loiseau \email{sylvain.loiseau@univ-paris13.fr}

}
