% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.emeld.R
\name{items.in.element}
\alias{items.in.element}
\title{Search the "item" elements in the nodes of an unit}
\usage{
items.in.element(elements, fields, languages)
}
\arguments{
\item{elements}{a set of node of the same unit (morphems, sentences...)}

\item{fields}{a character vector of field names.}

\item{languages}{a character vector of language names}
}
\value{
a data frame with as many columns as fields.
}
\description{
In EMELD vocabulary, most of the information is in "item" elements child of text,
word (=sentence), word or morphs elements.
These items have two attribute: @type for indicating the field
(part of speech, gloss, free translation...)
and @lang for indicating the language (either under scrutiny or used in the analysis).
}
