% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{kakabe}
\alias{kakabe}
\title{A corpus of 8 texts of the Kakabe language}
\format{A list with 4 slots}
\usage{
kakabe
}
\description{
The corpus is produced with the read.toolbox() function. It is a list of 4 slots representing four 
units: "texts" "sentences" "words" "morphems". Each slot contains a data frame, and each row
in the data.frame describe one occurrences of the corresponding unit.
}
\details{
\itemize{
  \item texts : a data frame of 8 units and 2 columns
  \item sentenes : a data frame of 552 units and 5 columns
  \item words : a data frame of 7381 units and 7 columns
  \item mophems numeric : a data frame of 8659 units and 9 columns
}
}
\references{
Data by Alexandra Vydrina <alexandra.vydrina@gmail.com>
}
\keyword{datasets}
