% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.specifications.R
\name{lift.specification}
\alias{lift.specification}
\alias{entry.field}
\alias{sense.field}
\alias{example.field}
\alias{relation.field}
\title{List of the available pieces of information for each entry (ie column in the entry table)}
\usage{
entry.field()

sense.field()

example.field()

relation.field()
}
\value{
a character vector of entries.

a character vector of entries.

a character vector of entries.

a character vector of entries.
}
\description{
List of the available pieces of information for each entry (ie column in the entry table)

List of the available pieces of information for each sense (ie column in the sense table)

List of the available pieces of information for each example (ie column in the example table)

List of the available pieces of information for each relation (ie column in the relation table)
}
