% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.toolbox.R
\name{line2field}
\alias{line2field}
\title{read and collapse multi-line fields into one line.
One can encouter:
    \\mb samuel -we  ta -li  -lo
    nefi    -mwii
where a field is actually on two lines.}
\usage{
line2field(lines)
}
\arguments{
\item{lines}{character vector : lines of the toolbox file}
}
\description{
read and collapse multi-line fields into one line.
One can encouter:
    \\mb samuel -we  ta -li  -lo
    nefi    -mwii
where a field is actually on two lines.
}
