% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.lift.R
\name{populate.table}
\alias{populate.table}
\title{Populate a data frame with pieces of information found in a node set according to a list of column to be crated.}
\usage{
populate.table(nodes, table, field.names, field.specs, vernacular.languages,
  analysis.languages)
}
\arguments{
\item{nodes}{a node set. Each node correspond to a row in the table. Nodes can be lexical entries, or exemples, or senses, or relations...}

\item{table}{a table containing some information to be augmented with new columns.}

\item{field.names}{character vector: the name of the field we are interested in. Each element in this vector
will result in at least one column (if an element denote an information given in verncular or analysis language,
and if several vernacular or analysis languages are declared (see the parameters "vernacular.languages" and
"analysis.languages", then this element will result in several column)}

\item{field.specs}{The specification (how to get information?) for all the available fields. See the page lift-format.}

\item{vernacular.languages}{character vector: the code of the vernacular languages used in the dictionary}

\item{analysis.languages}{character vector: the code of the vernacular languages used in the dictionary}
}
\value{
a data frame with the requested columns
}
\description{
Populate a data frame with pieces of information found in a node set according to a list of column to be crated.
}
