% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.CLDF.R
\name{read.CLDF}
\alias{read.CLDF}
\title{Read a set of csv files according to a metadata files following the CLDF format.}
\usage{
read.CLDF(directory, metadata_filename)
}
\arguments{
\item{directory}{path to the directory containing the CLDF bundle}

\item{metadata_filename}{name of the metadata file in the directory}
}
\value{
a list of data frame.
}
\description{
Read the metadata file in order to get the csv dialect specification, Module and Component information.
The CLDF name (for module and component) are recorded in the class attribute of the outer list and of the inner data.frame).
The metadata fragment describing a table is recorded in an "CLDF_metadata" attribute on each data frame.
}
