% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.emeld.R
\name{read.emeld.old}
\alias{read.emeld.old}
\title{Create a data frame importing an interlinearised corpus in ELMED XML.}
\usage{
read.emeld.old(file)
}
\arguments{
\item{file}{the uri of the XML document}
}
\value{
a data frame
}
\description{
The EMELD XML vocabulary has been proposed for the encoding of interlinear glosses.
It is used by the FieldWorks software (SIL FLEX) as an export format.
}
\examples{
path <- system.file("exampleData", "tuwariInterlinear.xml", package="RFlex")
corpus <- read.emeld.old(path)
}
\references{
Baden Hughes, Steven Bird and Catherine Bow \emph{Encoding and Presenting Interlinear Text Using XML Technologies}, http://www.aclweb.org/anthology/U03-1008
}
\seealso{
read.lift
}
