% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.lift.R
\name{read.lift.old}
\alias{read.lift.old}
\title{Parse a dictionary in XML LIFT (Lexicon Interchange FormaT) format and turn it into a data.frame}
\usage{
read.lift.old(file, language.code = "tww")
}
\arguments{
\item{file}{: a length-one character vector containing the path to a LIFT XML document.}

\item{language.code}{: the code of the language.}
}
\value{
a data.frame
}
\description{
Parse a dictionary in XML LIFT (Lexicon Interchange FormaT) format and turn it into a data.frame
}
\examples{
path <- system.file("exampleData", "tuwariDictionary.lift", package="RFlex")
dictionary <- read.lift.old(path, language.code="tww")
}
\references{
http://code.google.com/p/lift-standard
}
\seealso{
read.emeld
}
