% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.toolbox.R
\name{read.toolbox}
\alias{read.toolbox}
\title{Read a toolbox (SIL) text file and turn it into a list of data frame.}
\usage{
read.toolbox(path, add.unaligned = FALSE)
}
\arguments{
\item{path}{: the path to a toolbox text file.}

\item{add.unaligned}{: If TRUE: if any of the mb/ge/ps triplet does not have the same
number of morphems (resp. morphem glosses and part of speech),
an extra slot named 'unaligned' is added to the returned list giving information on all the unaligned triplets. If FALSE: print
a warning message only.}
}
\value{
a list with three slots "texts", "sentences" and "morphemes", each containing a data frame.
In these data frame, each row is an occurrence of the unit concerned.
}
\description{
The triplet of lines (mb, ge, ps) (containing reps. morphems, glosses
and part of speech) that do not have an equal number of white-space
separated item ('unaligned') are discarded.
}
\examples{
corpuspath <- system.file("exampleData", "tuwariToolbox.txt", package="interlineaR")
corpus <- read.toolbox(corpuspath)
}
\references{
https://software.sil.org/toolbox/
}
\seealso{
read.emeld
}
