% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFlex.R
\docType{package}
\name{RFlex-package}
\alias{RFlex}
\alias{RFlex-package}
\title{Parse XML document created by the Flex "export" function.}
\description{
Parse XML document created by the Flex "export" function.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.r-project.org}
}

}
