% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.toolbox.R
\name{tokenize.aligned.field.set}
\alias{tokenize.aligned.field.set}
\title{Tokenize a set of fields associated to an unit (the set of fields related to word segmentation for instance, or to morpheme segmentation).}
\usage{
tokenize.aligned.field.set(longformat, fields = c("mb", "ge", "ps"),
  masterfield = "mb", unit_id_name = "morphemes_id", included.in = NULL,
  included.in_id = "words_id")
}
\arguments{
\item{longformat}{the data table of the toolbox fields in long format (melted)}

\item{fields}{character vector: the aligned fields to tokenize (for instance, "mb", "ge" and "ps")}

\item{masterfield}{the master in the fields to be tokenized (for instance, "mb").}

\item{unit_id_name}{character vector (length-1): the name of the column to be created containing an ID for each token.}

\item{included.in}{character vector: an optional super-ordinated unit (for instance words for morphemes).}

\item{included.in_id}{character vector (length-1): the name of the column to be created containing an ID of the outer unit.}
}
\value{
a data frame with one token by row and as many columns as fields.
}
\description{
Set of fields have a "master field" (such as tx for words, or mb for morphemes), and other fields
associated with it (such as "ge" and "ps" for the morphemes master field.)
}
