% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{vatlongos}
\alias{vatlongos}
\title{A corpus of 10 texts of the Vatlongos (vtk) language}
\format{A list with 4 slots}
\usage{
vatlongos
}
\description{
The corpus is produced with the read.emeld() function. It is a list of 4 slots representing four 
units: "texts" "sentences" "words" "morphems". Each slot contains a data frame, and each row
in the data.frame describe one occurrences of the corresponding unit.
}
\details{
\itemize{
  \item texts : a data frame of 95 units and 5 columns ("text_id", "title.en", "title.abbreviation.en", "source.en", "comment.en")
  \item sentenes : a data frame of 3967 units and 6 columns ("text_id", "sentence_id", "segnum.en", "gls.en", "lit.en", "note.en")
  \item words : a data frame of 52983 units and 6 columns ("text_id" "sentence_id" "word_id" "txt.tvk" "gls.en" "pos.en")
  \item mophems numeric : a data frame of 56354 units and 10 columns ("text_id" "sentence_id" "word_id" "morphem_id" "type" "txt.tvk" "cf.tvk" "gls.en" "msa.en" "hn.en" )
}

See the vignette vatlongos for Case study based on this corpus.
}
\references{
Eleanor Ridge <Eleanor_Ridge@soas.ac.uk>
}
\keyword{datasets}
