% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.CLDF.R
\name{create_CLDF_metadata}
\alias{create_CLDF_metadata}
\title{Create the CLDF metadata for a list of data frame representing a CLDF component
(Supported: Dictionary, interlinear glossed text corpus)}
\usage{
create_CLDF_metadata(l)
}
\arguments{
\item{l}{the list of data.frame

This function tries to generate the metadata tree according to the following heuristic:
\itemize{
  \item if the list or some of the inner tables have an attribute "CLDF_metadata", the content of
  this attribute is supposed to be a tree of lists expressing the metadata for this object and it
  is aggregated to the metadata.
  \item if the outer list has CLDF Module name in its "class" attributes,
  it will be used for generating the correct Module metadata.
  If some of the inner table have a CLDF Component name in their class attribute,
  it will also be used for generating the correct metadata.
  For instance, a list with the value "Dictionary" in its class attribute, containing tables with
  "EntryTable" and "SenseTable" in their class attribues, will result in a metadata file describing
  a Dictionary module with its two mandatory components.
  \item In last resort, if no information is found about the table, the names of their slot in the outer list will be
  used as url and as component name in the metadata file.
}}
}
\value{
a tree of list containing the metadata. This tree is ready to be serialized using jsonlite::toJSON.
}
\description{
Metadata are expressed, in R, as a tree of lists. This tree can be converted to JSON using jsonlite::toJSON. For instance,
\verb{
list(
  format=list(
    x="a",
    y="b"
  ),
  version=list(
    "1",
    "b"
  )
)
}
will result in the following JSON fragment:
\verb{
{
	"format": {
		"x": "a",
		"y": "b"
	},
	"version": [
		"1",
		"b"
		]
}
}
}
\references{
http://cldf.clld.org https://github.com/cldf/cldf/
}
