% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.lift.R
\name{read.lift}
\alias{read.lift}
\title{Parse a dictionary in XML LIFT (Lexicon Interchange FormaT) format and turn it into a data.frame}
\usage{
read.lift(file)
}
\arguments{
\item{file}{: a length-one character vector containing the path to a LIFT XML document.}
}
\value{
a data.frame
}
\description{
Parse a dictionary in XML LIFT (Lexicon Interchange FormaT) format and turn it into a data.frame
}
\examples{
path <- system.file("exampleData", "tuwariDictionary.lift", package="RFlex")
dictionary <- read.lift(path)
}
\references{
http://code.google.com/p/lift-standard
}
\seealso{
read.emeld
}
